Oh, core profile. Why do you mock me so?
Turns out that I've been having issues not because I was doing
something wrong math-wise, but because I was doing something
wrong GL-wise. I wasn't properly creating VAOs for the chunk meshes,
and for various reasons it "worked" until I started rendering
alpha-blended faces as well.

Now it clicked together more how VBOs and VAOs are related, and I
made some changes to the mesh code: There is no renderable or renderobj
anymore, a mesh is a VAO plus a data VBO plus an optional index VBO.

Setting the material to use is separate from drawing with a mesh, but
a mesh needs to be set to use the material that is set to use it. :D

I think I need a break.