Change license to public domain (unlicense)
The only difference compared to the ISC license I originally used is
that public domain does not require attribution. Attribution is a
meaningless clause anyway that only leads to a bunch of legalese stuck
in the manual somewhere that no one reads or cares about. Feel free to
use this code however and wherever you please.

Note that any source files with a different license in their header
remain under that license, the public domain dedication only applies to
those files that do not explicitly declare themselves under a different
license.