thinking about chunks
Simplified the chunk handling quite a bit with the realisation that
I can just use the dirty bit to track whether a chunk needs to be
remeshed. This way, I can just set the dirty flag and each tick, I check
the dirty flags and load up to five (?) dirty chunks.

Simple.